I'm thinking about doing regular benches, something like every week bench master vs last week's master.

We did 8.19 vs 8.18 recently https://github.com/coq/coq/pull/18428 so V8.20+alpha is a reasonable base to start.