@SkySkimmer 2024-01-12
The `coq` slowdown (ie test suite) is probably added tests in https://github.com/coq/coq/pull/17933

Testing locally:
- master: 2m30s (150s) to build the vo, 21s for coqchk
- with #17933 reverted: 29s to build the vo, 29s or coqchk

ie master takes +120s
assuming my machine is a bit faster than the bench machines that matches the +150s in the bench